Add files via upload
This practical will focus on sending electronic mail using the SMTP protocol
Despite the various social chat and connect applications that are available today, e-mail still
remains a popular communication tool. Wouldn’t it be interesting to understand how some
of the protocols behind this essential communication method work? This week you are being
tasked to look at SMTP and build an SMTP client.
Create a GUI based Java application which will act as a mail client that sends e-mail. When
the application starts up the user must be presented with a user interface. The interface
should allow the user to enter the host-name and port number to which they wish to connect.
Furthermore, the GUI should contain textfields (with associated labels) for sender name,
recipient name and a text area for the contents of the email. Finally a send button on the
interface will allow the message to be sent. Refer to the SMTP protocol for information on
how this is done.
After the user has clicked send, the user will be notified on the status of the sent email. Any
errors which occur must be displayed to the user, in a user friendly manner. Testing of the
mail client will need to be setup. For a mail server which can be used for testing, refer to
SMTP Bucket or Papercut (Papercut.Smtp.Setup.exe can be downloaded here). Being able
to use SMTP Bucket or Papercut is not a requirement!
Please note: You may NOT use the JavaMail API (javax.mail package) to complete your
practical. Failure to code the protocol from scratch will result in receiving 0 for your practical.